+++ Cargo.toml
@@ -172,7 +172,7 @@ hkdf = "0.12.3"
 
 # webgpu
 raw-window-handle = "0.5.0"
-wgpu-core = "=0.18"
+wgpu-core = { version = "=0.18", features = ["raw-window-handle"] }
 wgpu-types = "=0.18"
 wgpu-hal = "=0.18"
