+++ ext/webgpu/surface.rs
@@ -11,17 +11,6 @@ use std::borrow::Cow;
 use std::rc::Rc;
 use wgpu_types::SurfaceStatus;
 
-deno_core::extension!(
-  deno_webgpu_surface,
-  deps = [deno_webidl, deno_web, deno_webgpu],
-  ops = [
-    op_webgpu_surface_configure,
-    op_webgpu_surface_get_current_texture,
-    op_webgpu_surface_present,
-  ],
-  esm = ["02_surface.js"],
-);
-
 pub struct WebGpuSurface(pub crate::Instance, pub wgpu_core::id::SurfaceId);
 impl Resource for WebGpuSurface {
   fn name(&self) -> Cow<str> {
